Revert "More Baubles support"
This reverts commit e8feb577b706000e38e0c22758bdb0336513346e.
