+++ resources/assets/enderio/lang/en_US.lang
@@ -559,7 +559,7 @@ item.itemMagnet.name=Electromagnet
 item.itemMagnet.tooltip.detailed.line1=Pulls nearby item towards you
 item.itemMagnet.tooltip.detailed.line2=Shift-R-Click to enable / disable
 item.itemMagnet.tooltip.detailed.line3=Must be powered, active and in the
-item.itemMagnet.tooltip.detailed.line4=hot bar (or a Baubles slot) to function
+item.itemMagnet.tooltip.detailed.line4=hot bar to function
 
 item.enderio.itemConduitProbe.name=Conduit Probe
 item.enderio.itemConduitProbe.tooltip.detailed.line1=In probe mode R-Click on a conduit
